+++ sdk/communication/Azure.Communication.JobRouter/src/Models/ExceptionTrigger.cs
@@ -1,23 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System.Text.Json;
-using System.Text.Json.Serialization;
-using Azure.Core;
-
 namespace Azure.Communication.JobRouter
 {
-    public abstract partial class ExceptionTrigger : IUtf8JsonSerializable
+    public abstract partial class ExceptionTrigger
     {
         /// <summary> The type discriminator describing a sub-type of ExceptionTrigger. </summary>
         public ExceptionTriggerKind Kind { get; protected set; }
-
-        void IUtf8JsonSerializable.Write(Utf8JsonWriter writer)
-        {
-            writer.WriteStartObject();
-            writer.WritePropertyName("kind"u8);
-            writer.WriteStringValue(Kind.ToString());
-            writer.WriteEndObject();
-        }
     }
 }