+++ sdk/cosmosdb/Azure.ResourceManager.CosmosDB/src/Custom/Models/MongoDBRoleDefinitionData.cs
@@ -10,10 +10,10 @@ using Azure.ResourceManager.CosmosDB.Models;
 
 namespace Azure.ResourceManager.CosmosDB
 {
+    [CodeGenSerialization(nameof(RoleDefinitionType), DeserializationValueHook = nameof(ReadRoleDefinitionType))]
     public partial class MongoDBRoleDefinitionData
     {
         /// <summary> Indicates whether the Role Definition was built-in or user created. </summary>
-        [CodeGenMemberSerializationHooks(DeserializationValueHook = nameof(ReadRoleDefinitionType))]
         public MongoDBRoleDefinitionType? RoleDefinitionType { get; set; }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]