+++ sdk/monitor/Azure.Monitor.OpenTelemetry.AspNetCore/src/Vendoring/OpenTelemetry.ResourceDetectors.Azure/ResourceAttributeConstants.cs
@@ -1,18 +1,5 @@
-// <copyright file="ResourceAttributeConstants.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// </copyright>
+// SPDX-License-Identifier: Apache-2.0
 
 namespace OpenTelemetry.ResourceDetectors.Azure;
 
@@ -35,8 +22,14 @@ internal sealed class ResourceAttributeConstants
     internal const string AppServiceSlotNameEnvVar = "WEBSITE_SLOT_NAME";
     internal const string AppServiceStampNameEnvVar = "WEBSITE_HOME_STAMPNAME";
 
+    // Azure Container Apps environment variables
+    internal const string AzureContainerAppsNameEnvVar = "CONTAINER_APP_NAME";
+    internal const string AzureContainerAppsReplicaNameEnvVar = "CONTAINER_APP_REPLICA_NAME";
+    internal const string AzureContainerAppsRevisionEnvVar = "CONTAINER_APP_REVISION";
+
     // Azure resource attributes constant values
     internal const string AzureAppServicePlatformValue = "azure_app_service";
     internal const string AzureCloudProviderValue = "azure";
     internal const string AzureVmCloudPlatformValue = "azure_vm";
+    internal const string AzureContainerAppsPlatformValue = "azure_container_apps";
 }